pyproject.toml
[project]
name = "mcp-model-context-protocol"
version = "0.1.0"
description = "Project demonstrating MCP (Model Context Protocol) communication between apps"
authors = []
dependencies = [
    "fastapi==0.109.2",
    "uvicorn==0.27.1",
    "requests==2.31.0",
    "python-dotenv==1.0.1"
]
requires-python = ">=3.12"

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "pytest"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server", "app_a", "app_b"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
